# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Card:
    properties:
      _id:
        type: string
      card_holder:
        type: string
      color:
        type: string
      currency:
        type: string
      last_digit:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  requests.Card:
    properties:
      card_holder:
        type: string
      color:
        type: string
      currency:
        type: string
      last_digit:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - card_holder
    - color
    - currency
    - last_digit
    - name
    - type
    type: object
  requests.CardUpdate:
    properties:
      card_holder:
        type: string
      color:
        type: string
      currency:
        type: string
      id:
        type: string
      last_digit:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - id
    type: object
...
  requests.FavouriteInvestingCreate:
    properties:
      market:
        type: string
      priority:
        type: integer
      symbol:
        type: string
      type:
        type: string
    required:
    - market
    - priority
    - symbol
    - type
    type: object
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  responses.CardStats:
    properties:
      subscription_stats:
        $ref: '#/definitions/responses.CardSubscriptionStatistics'
      transaction_stats:
        $ref: '#/definitions/responses.TransactionTotal'
    type: object
  responses.CardSubscriptionStatistics:
    properties:
      currency:
        type: string
      total_monthly_payment:
        type: number
      total_payment:
        type: number
    type: object
...
  responses.FavouriteInvesting:
    properties:
      _id:
        type: string
      investing:
        $ref: '#/definitions/responses.Investing'
      investing_id:
        $ref: '#/definitions/responses.FavouriteInvestingID'
      priority:
        type: integer
      user_id:
        type: string
    type: object
  responses.FavouriteInvestingID:
    properties:
      market:
        type: string
      symbol:
        type: string
      type:
        type: string
    type: object
  responses.Investing:
    properties:
      currency:
        type: string
      price:
        type: number
    type: object
...
  responses.TransactionTotal:
    properties:
      currency:
        type: string
      total_transaction:
        type: number
    type: object
...
host: rocky-reaches-65250.herokuapp.com
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Kantan.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://rocky-reaches-65250.herokuapp.com/terms/
  title: Kantan Investment Manager API
  version: "1.0"
paths:
...
  /card:
    delete:
      consumes:
      - application/json
      description: Deletes card by id
      parameters:
      - description: ID
        in: body
        name: ID
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete card by card id
      tags:
      - card
    get:
      consumes:
      - application/json
      description: Returns cards by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Card'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Cards by User ID
      tags:
      - card
    post:
      consumes:
      - application/json
      description: Creates card
      parameters:
      - description: Card Create
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/requests.Card'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Card'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Card
      tags:
      - card
    put:
      consumes:
      - application/json
      description: Updates card
      parameters:
      - description: Card Update
        in: body
        name: cardupdate
        required: true
        schema:
          $ref: '#/definitions/requests.CardUpdate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Card'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: Unauthorized update
          schema:
            type: string
        "404":
          description: Couldn't find user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Card
      tags:
      - card
  /card/all:
    delete:
      consumes:
      - application/json
      description: Deletes all cards by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete all cards by user id
      tags:
      - card
  /card/stats:
    get:
      consumes:
      - application/json
      description: Returns card statistics
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.CardStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Card Statistics by User ID & Card ID
      tags:
      - card
...
  /watchlist:
    delete:
      consumes:
      - application/json
      description: Deletes favourite investing by id
      parameters:
      - description: ID
        in: body
        name: ID
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete favourite investing by favourite investing id
      tags:
      - favouriteinvesting
    get:
      consumes:
      - application/json
      description: Returns favourite investings by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.FavouriteInvesting'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Favourite Investings by User ID
      tags:
      - favouriteinvesting
    post:
      consumes:
      - application/json
      description: Creates favourite investing
      parameters:
      - description: Favourite Investing Create
        in: body
        name: favouriteinvestingcreate
        required: true
        schema:
          $ref: '#/definitions/requests.FavouriteInvestingCreate'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Favourite Investing
      tags:
      - favouriteinvesting
  /watchlist/all:
    delete:
      consumes:
      - application/json
      description: Deletes favourite investing by user id
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete all favourite investings by user id
      tags:
      - favouriteinvesting
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"